+++ CHANGELOG.md
@@ -7,6 +7,10 @@
  - Added support for setting `gh_pages_visibility` in repositories. ([#474](https://github.com/eclipse-csi/otterdog/pull/474))
  - Added support for setting `members_can_create_private_pages` in organization settings. ([#474](https://github.com/eclipse-csi/otterdog/pull/474))
 
+### Changed
+
+ - Make the `sync-check` always succeed to not prevent PRs from being merged due to unrelated changes in the live settings.
+
 ### Fixed
 
  - Properly escape string values when passing them to a rich Console. ([#440](https://github.com/eclipse-csi/otterdog/issues/440))