+++ CHANGELOG.md
@@ -8,7 +8,16 @@
 
 ### Fixed
 
+ - Fixed GitHub Page updates due missing required fields ([#450](https://github.com/eclipse-csi/otterdog/pull/455))
  - Properly escape string values when passing them to a rich Console. ([#440](https://github.com/eclipse-csi/otterdog/issues/440))
+
+### Changed
+
+ - Contributing documentation to include instructions for Otterdog Web App
+
+### Added
+
+ - Add Documentation for Otterdog Web App deployment
  - Add support for marking organizations as archived and to ignore them. ([#463](https://github.com/eclipse-csi/otterdog/issues/463))
  - Add new check-status CLI command. This outputs a json file of with sync, validation, and archival status of organizations. ([#457](https://github.com/eclipse-csi/otterdog/issues/457))
  - Fix regression of live settings and expected code scanning default setup ([#458](https://github.com/eclipse-csi/otterdog/issues/458))